+++ packages/qulacs/tests/qulacs/estimator/test_estimator.py
@@ -49,7 +49,7 @@ from quri_parts.qulacs.estimator import (
 
 
 def create_vector(qubit_count: int, bits: int) -> StateVectorType:
-    vector: StateVectorType = np.zeros(2**qubit_count, dtype=np.cfloat)
+    vector: StateVectorType = np.zeros(2**qubit_count, dtype=np.complex128)
     vector[bits] = 1.0
     return vector
